Finished CheckboxComponent Registry
CheckboxGroup component was added, the CheckBox component is going to be assess for usages, however, the current changes are in Component Registry and added to Types with seemless integration
