+++ frontend/src/config.json
@@ -1,5 +1,5 @@
 {
-  "max_courses": 4,
+  "max_courses": 5,
   "loadData" : true,
   "saveData" : true,
   "test_login": true,
