+++ frontend/src/utils/ComponentRegistry.ts
@@ -1,9 +1,11 @@
 import AccordionWrapper from "../components/SyllabusComponents/ComponentWrappers/AccordionWrapper";
 import ColumnWrapper from "../components/SyllabusComponents/ComponentWrappers/ColumnWrapper";
 import RowWrapper from "../components/SyllabusComponents/ComponentWrappers/RowWarapper";
+import CheckboxGroupWrapper from "../components/SyllabusComponents/ComponentWrappers/CheckboxGroupWrapper";
 
 export const ComponentRegistry = {
     Accordion: AccordionWrapper,
     Column: ColumnWrapper,
-    Row: RowWrapper
+    Row: RowWrapper,
+    CheckBoxGroup: CheckboxGroupWrapper
 }
